chore: convert main module to use babel
