+++ lib/config.js
@@ -1,0 +1,28 @@
+'use babel';
+
+export default {
+  useCustomUrlTemplateIfStandardRemotesFail : {
+    type: 'boolean',
+    default: false
+  },
+  customCommitUrlTemplateString : {
+    type: 'string',
+    default: 'Example -> https://github.com/<%- project %>/<%- repo %>/commit/<%- revision %>'
+  },
+  columnWidth : {
+    type: 'integer',
+    default: 210
+  },
+  dateFormatString : {
+    type: 'string',
+    default: 'YYYY-MM-DD'
+  },
+  ignoreWhiteSpaceDiffs : {
+    type: 'boolean',
+    default: false
+  },
+  showOnlyLastNames : {
+    type: 'boolean',
+    default: true
+  },
+}
