+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "git-blame",
-  "main": "./lib/git-blame",
+  "main": "./lib/index",
   "version": "0.4.12",
   "description": "Toggle git-blame annotations in the gutter of atom editor.",
   "activationCommands": {
